+++ tests/testthat/test-matrix.R
@@ -14,21 +14,6 @@ expect_true(getValidity(getClassDef("CsparseMatrix"))(x))
 expect_true(.Call(arules:::R_valid_ngCMatrix, x))
 expect_true(.Call(arules:::R_valid_ngCMatrix, e))
 
-## t
-expect_identical(.Call(arules:::R_transpose_ngCMatrix, x), t(x))
-expect_identical(.Call(arules:::R_transpose_ngCMatrix, e), t(e))
-
-## column/row subset (index can only be integer now)
-s <- as.integer(c(1,1,3,4))
-expect_equal(x[,s], .Call(arules:::R_colSubset_ngCMatrix, x, s))
-
-#
-expect_identical(.Call(arules:::R_colSubset_ngCMatrix, e, integer()), e)
-expect_equal(x[s,], .Call(arules:::R_rowSubset_ngCMatrix, x, s))
-
-#
-expect_identical(.Call(arules:::R_rowSubset_ngCMatrix, e, integer()), e)
-
 ## reorder
 expect_equal(.Call(arules:::R_recode_ngCMatrix, x, 6:1), x[6:1,])
 expect_equal(.Call(arules:::R_recode_ngCMatrix, e, integer()), e)
@@ -48,14 +33,6 @@ expect_identical(.Call(arules:::R_or_ngCMatrix, x, x),
   as(x | x, "nsparseMatrix"))
 expect_identical(.Call(arules:::R_or_ngCMatrix, e, e), 
   as(e | e, "nsparseMatrix"))
-
-## row sums
-expect_equal(.Call(arules:::R_rowSums_ngCMatrix, x), rowSums(x))
-expect_equal(.Call(arules:::R_rowSums_ngCMatrix, e), rowSums(e))
-
-## column sums
-expect_equal(colSums(x), .Call(arules:::R_colSums_ngCMatrix, x))
-expect_equal(.Call(arules:::R_colSums_ngCMatrix, e), colSums(e))
 
 ## crossprod and tcrossprod
 expect_equal(.Call(arules:::R_crosstab_ngCMatrix, x, NULL, TRUE),