+++ NEWS.md
@@ -4,6 +4,14 @@
 ## Changes
 * Updated tests for testthat edition 3.
 
+## Internal Changes
+* Removed internal ngCMatrix subsetting code which has issues with R-devel.
+  We use now subsetting provided by package Matrix which is almost as fast.
+* Removed internal code for rowSums and colSums for ngCMatrix. 
+* Internal code for t for ngCMatrix is now only used internally. 
+
+
+
 ## Bugfixes
 * Fixed 'Error in .basicRuleMeasure(..) unused argument 
   (complement = TRUE)' reported by bachnguyen-tomo.